+++ Utopia/UTexture.cpp
@@ -106,7 +106,7 @@ void UTexture::setAnisotropyLevel(int level)
         if (level < 1)
             level = 1;
 
-    glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAX_ANISOTROPY_EXT, level);
+    glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAX_ANISOTROPY_EXT, (GLfloat)level);
 }
 
 void utopia::UTexture::updateTextureParameteri(void(*parametriSetMethod)(void))
